+++ src/util/metricsAggregator.js
@@ -2,6 +2,7 @@
 const cluster = require('cluster');
 const logger = require('../logger');
 const { Worker, isMainThread } = require('worker_threads');
+const v8 = require('v8');
 
 const MESSAGE_TYPES = {
   GET_METRICS_REQ: 'rudder-transformer:getMetricsReq',
@@ -94,7 +95,7 @@ class MetricsAggregator {
         const metrics = await this.prometheusInstance.prometheusRegistry.getMetricsAsJSON();
         cluster.worker.send({
           type: MESSAGE_TYPES.GET_METRICS_RES,
-          metrics: JSON.stringify(metrics),
+          metrics,
           requestId: message.requestId,
         });
       } catch (error) {
@@ -297,7 +298,7 @@ class MetricsAggregator {
       this.resetAggregator(true);
       return;
     }
-    this.metricsBuffer.push(JSON.parse(message.metrics));
+    this.metricsBuffer.push(message.metrics);
     this.pendingMetricRequests--;
     if (this.pendingMetricRequests === 0) {
       this.aggregateMetricsInWorkerThread();