+++ test/integrations/destinations/snapchat_conversion/router/dataV3.ts
@@ -470,6 +470,24 @@ export const dataV3: RouterTestData[] = [
                       {
                         action_source: 'MOBILE_APP',
                         app_data: {
+                          extinfo: [
+                            'i2',
+                            'com.rudderlabs.javascript',
+                            '1.0.0',
+                            '1.0.0',
+                            '14.4.1',
+                            'AOSP on IA Emulator',
+                            'en-US',
+                            '',
+                            '',
+                            '',
+                            '',
+                            '2',
+                            '',
+                            '',
+                            '',
+                            '',
+                          ],
                           app_id: 'dhfeih44f',
                         },
                         custom_data: {
@@ -489,6 +507,24 @@ export const dataV3: RouterTestData[] = [
                       {
                         action_source: 'MOBILE_APP',
                         app_data: {
+                          extinfo: [
+                            'i2',
+                            'com.rudderlabs.javascript',
+                            '1.0.0',
+                            '1.0.0',
+                            '14.4.1',
+                            '',
+                            'en-US',
+                            '',
+                            '',
+                            '',
+                            '',
+                            '2',
+                            '',
+                            '',
+                            '',
+                            '',
+                          ],
                           app_id: 'dhfeih44f',
                         },
                         custom_data: {